requirements.txt
accelerate
addict
aiohttp
attrdict
binpacking
charset_normalizer
cpm_kernels
dacite
datasets>=3.0,<3.4
einops
fastapi
gradio>=3.40.0
importlib_metadata
jieba
matplotlib
modelscope>=1.23
nltk
numpy<2.0
openai
oss2
pandas
peft>=0.11,<0.16
pillow
requests
rouge
safetensors
scipy
sentencepiece
simplejson>=3.3.0
sortedcontainers>=1.5.9
tensorboard
tiktoken
tqdm
transformers>=4.33,<4.53
transformers_stream_generator
trl>=0.15,<0.18
uvicorn
zstandard
timeout_decorator
autopep8
vllm==0.7.3
transformers==4.51.1
qwen_vl_utils==0.0.11
pylatexenc
trl==0.17.0
